+++ CHANGELOG.md
@@ -13,6 +13,11 @@ Changed:
 
  - Use more correct "Alpaca Mount" terminology instead of "ASCOM Remote" #38
 
+Fixed:
+
+ - "Reset Settings" no longer deletes your saved settings. If you want to
+	reset your saved settings, please re-save your settings. #40
+
 ## v2.1.0 - 2021-07-31
 
 Added: